Refactor: change DTO name, RecordMember to RecordDTO
